消除编译warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@29191 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/spt/usrdef/sptUsrCmdCommon.cpp
@@ -113,6 +113,7 @@ namespace engine
       INT32 rc = SDB_OK ;
       ossCmdRunner runner ;
       OSSHANDLE processHandle = (OSSHANDLE)0 ;
+
 #if defined( _LINUX )
       struct sigaction    ignore ;
       struct sigaction    savechild ;
@@ -181,7 +182,7 @@ namespace engine
          OSSPID childPid = runner.getPID() ;
          if ( 0 != processHandle )
          {
-            UINT32 times = 0 ;
+            INT64 times = 0 ;
             while ( times < timeout )
             {
                if ( ossIsProcessRunning( childPid ) )
